ci: increase postgres max_connections limit to 200
There is no available env var to set a custom max_connections value. In
order to do this, we need to modify the database configuration itself
which can be done by running the 'alter system set' psql command.

Postgres needs to be restarted in order to apply the new configuration.

Value for the max_connections ranges from 1..262143. Setting the value
to 200 for now as our integration tests only require around 100.